Fixed wrong data types and corrected import algorithm

+++ src/Graphics/Model.hpp
@@ -14,20 +14,21 @@ namespace DevaFramework
 	public:
 
 		DEVA_GRAPHICS_API static Model fromFile(const std::string &filename);
+		/**TODO: NOT FULLY IMPLEMENTED!!!!!!*/
 		DEVA_GRAPHICS_API static std::vector<char> exportBinary(const Model &model);
 
 	private:
 		
 		VBO vbo;
 
-		std::vector<unsigned int> index_arr;
+		std::vector<uint16_t> index_arr;
 
 	public:
 		DEVA_GRAPHICS_API Model();
-		DEVA_GRAPHICS_API Model(const VBO &vbo, const std::vector<unsigned int> &index_arr);
+		DEVA_GRAPHICS_API Model(const VBO &vbo, const std::vector<uint16_t> &index_arr);
 
 		DEVA_GRAPHICS_API const VBO& getVBO() const;
-		DEVA_GRAPHICS_API const std::vector<unsigned int>& getIndexArray() const;
+		DEVA_GRAPHICS_API const std::vector<uint16_t>& getIndexArray() const;
 
 		DEVA_GRAPHICS_API std::string str();
 	};
